This is a simple explanation of Diffie Hellman
Diffie Hellman has a been a very useful method these days for
communication over the internet. It proves to be used for creating one
privatekey that two sides communication with each other can have. This
is done without having to exchange any sensitive data in public more
than agreeing on some numbers that are used to calculate a public key
and a secret key that could be exchanged. With having these numbers, an
attacker can very hardly guess what the calculated secret key is.